requirements.txt
django
django-crispy-forms
celery
requests
django-filter
djangorestframework
drf-yasg
djangorestframework-xml
djangorestframework-csv
djangorestframework-simplejwt
gunicorn
uwsgi
beautifulsoup4
lxml
psycopg2-binary
django-environ
django-admin-rangefilter

# dev
django-debug-toolbar
django-extensions
ipython
flower
django-silk

# flake8
flake8
flake8-print
flake8-builtins
flake8-import-order
flake8-debugger

# pytest
pytest
pytest-django
pytest-cov
pytest-mock
